Safari-611.1.10/LayoutTests/http/tests/subresource-integrity/sri-enabled-with-setting-expected.txt

CONSOLE MESSAGE: Cannot load script http://127.0.0.1:8000/subresource-integrity/resources/non-matching-digest.js. Failed integrity metadata check. Content length: 25, Expected content length: 25, Expected metadata: sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead
CONSOLE MESSAGE: Cannot load stylesheet http://127.0.0.1:8000/subresource-integrity/resources/style.css. Failed integrity metadata check. Content length: 37, Expected content length: 37, Expected metadata: sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead

PASS Test that Subresource Integrity's 'integrity' property is exposed on the <link> element.
PASS Test that Subresource Integrity's 'integrity' property is exposed on the <script> element.
PASS Script: Same-origin with incorrect hash, should block evaluation.
PASS Style: Same-origin with incorrect hash, should block evaluation.
